add server subcommand
