+++ cmd/cpace-machine/main.go
@@ -18,6 +18,7 @@ var subcmds = map[string]func(args ...string){
 	"send":    send,
 	"receive": receive,
 	"pipe":    pipe,
+	"server":  server,
 }
 
 var (
